requirements.txt
vk-api==11.10.0
groq==0.31.1
python-dotenv==1.1.1
requests==2.32.5
aiohttp==3.12.15
pytest==8.3.4
pytest-asyncio==0.24.0
psutil==6.1.0